Using static casting & deleting hidden files!

+++ apps/RandomnessGenerator.cpp
@@ -1,21 +1,17 @@
 #include "RandomnessGenerator.hpp"
 
-float RandomnessGenerator::getRandomStep() const
-{
-    // Seeding the random number generator with time(null) to get a different seed each time this function is called
-    return {((float)rand()) / (float)RAND_MAX};
-}
+// Simple random number generator function, generates a float between 0.0 and 1.0
+float RandomnessGenerator::getRandomStep() const { return {((float)rand()) / (float)RAND_MAX}; }
 
+// Returns a Point object that has randomized x,y and z coordinates after converting from randomized spherical coordinates
 Point RandomnessGenerator::getRandomPoint()
 {
-    // Seeding the random number generator with time(null) to get a different seed each time this function is called
-
     Point point; // Instance of the Point struct to return with the random coordinates
 
     // Getting random values for spherical coordinates transformation parameters
-    float u = ((float)rand()) / (float)RAND_MAX;
-    float v = ((float)rand()) / (float)RAND_MAX;
-    float r = cbrt(((float)rand()) / (float)RAND_MAX); // Cubic root to prevent clumping in the center
+    float u = static_cast<float>(((float)rand()) / (float)RAND_MAX);
+    float v = static_cast<float>(((float)rand()) / (float)RAND_MAX);
+    float r = static_cast<float>(cbrt(((float)rand()) / (float)RAND_MAX)); // Cubic root to prevent clumping in the center
 
     // Transformation equations
     float theta = 2 * M_PI * u;
@@ -32,6 +28,7 @@ Point RandomnessGenerator::getRandomPoint()
     return point;
 }
 
+// A helper function to generate a csv file to use in plotting
 void RandomnessGenerator::exportSamplingPlot(Point point)
 {
     // For streaming out my output in a log file
